+++ packages/ai-native/src/browser/index.ts
@@ -11,13 +11,13 @@ import { IAINativeService } from '../common/index';
 
 import { AINativeBrowserContribution } from './ai-core.contribution';
 import { AINativeService } from './ai-native.service';
-import { InlineChatFeatureRegistry } from './inline-chat-widget/inline-chat.feature.registry';
-import { AIInlineChatService } from './inline-chat-widget/inline-chat.service';
-import { IAINativeCoreContribution, IInlineChatFeatureRegistry } from './types';
+import { AINativeCoreContribution, IInlineChatFeatureRegistry } from './types';
+import { InlineChatFeatureRegistry } from './widget/inline-chat/inline-chat.feature.registry';
+import { AIInlineChatService } from './widget/inline-chat/inline-chat.service';
 
 @Injectable()
 export class AINativeModule extends BrowserModule {
-  contributionProvider = IAINativeCoreContribution;
+  contributionProvider = AINativeCoreContribution;
   providers: Provider[] = [
     AINativeBrowserContribution,
     {
