+++ packages/ai-native/src/browser/widget/inline-chat/inline-chat.module.less
