+++ packages/ai-native/src/browser/widget/inline-diff/inline-diff-widget.module.less
